requirements.txt
torch>=2.4.0
torchvision>=0.19.0
transformers>=4.40.0
accelerate>=1.10.0
safetensors>=0.6.0
tokenizers>=0.22.0
qwen-vl-utils>=0.0.14
av>=16.0.0
decord>=0.6.0
opencv-python-headless>=4.5.0
pillow>=11.0.0
imageio>=2.37.0
numpy>=1.24.0
scipy>=1.15.0
pandas>=2.0.0
tqdm>=4.67.0
pyyaml>=6.0.0
einops>=0.8.0
